+++ cli/aws_orbit/remote_files/cdk/lambda_sources/cognito_post_authentication/k8s_manage.py
@@ -6,7 +6,8 @@ import time
 from typing import Any, Dict, List, Optional, cast
 
 import boto3
-from kubernetes import client, config
+from kubernetes import client, config, dynamic
+from kubernetes.client import api_client
 from kubernetes.client.rest import ApiException
 
 logger = logging.getLogger()
@@ -16,11 +17,15 @@ ORBIT_ENV = os.environ.get("ORBIT_ENV")
 ACCOUNT_ID = os.environ.get("ACCOUNT_ID")
 REGION = os.environ.get("REGION")
 ROLE_PREFIX = os.environ.get("ROLE_PREFIX")
-
+ORBIT_API_VERSION = os.environ.get("ORBIT_API_VERSION", "v1")
+ORBIT_API_GROUP = os.environ.get("ORBIT_API_GROUP", "orbit.aws")
+ORBIT_SYSTEM_NAMESPACE = os.environ.get("ORBIT_SYSTEM_NAMESPACE", "orbit-system")
+ORBIT_STATE_PATH = os.environ.get("ORBIT_STATE_PATH", "/state")
+USERSPACE_CR_KIND = "UserSpace"
 KUBECONFIG_PATH = "/tmp/.kubeconfig"
 
-ssm = boto3.client("ssm")
 
+ssm = boto3.client("ssm")
 context = ssm.get_parameter(Name=f"/orbit/{ORBIT_ENV}/context")
 EFS_FS_ID = json.loads(context.get("Parameter").get("Value")).get("SharedEfsFsId")
 
@@ -29,8 +34,11 @@ def handler(event: Dict[str, Any], context: Optional[Dict[str, Any]]) -> Any:
     create_kubeconfig()
 
     api_CoreV1 = client.CoreV1Api()
+    userspace_dc = dynamic.DynamicClient(client=api_client.ApiClient()).resources.get(
+        group=ORBIT_API_GROUP, api_version=ORBIT_API_VERSION, kind=USERSPACE_CR_KIND
+    )
 
-    manage_user_namespace(event, api_CoreV1)
+    manage_user_namespace(event, api_CoreV1, userspace_dc)
 
 
 def run_command(cmd: str) -> None:
@@ -80,13 +88,49 @@ def create_user_efs_endpoint(user: str, team_name: str) -> Dict[str, Any]:
     )
 
 
+def create_userspace(
+    userspace_dc: dynamic.DynamicClient,
+    name: str,
+    env: str,
+    space: str,
+    team: str,
+    user: str,
+    user_efsapid: str,
+    user_email: str,
+    owner_reference: Optional[Dict[str, str]] = None,
+    labels: Optional[Dict[str, str]] = None,
+    annnotations: Optional[Dict[str, str]] = None,
+) -> None:
+    userspace: Dict[str, Any] = {
+        "apiVersion": f"{ORBIT_API_GROUP}/{ORBIT_API_VERSION}",
+        "kind": USERSPACE_CR_KIND,
+        "metadata": {"name": name, "labels": labels, "annotations": annnotations, "namespace": name},
+        "spec": {
+            "env": env,
+            "space": space,
+            "team": team,
+            "user": user,
+            "userEfsApId": user_efsapid,
+            "userEmail": user_email,
+        },
+    }
+    if owner_reference is not None:
+        userspace["metadata"]["ownerReferences"] = [owner_reference]
+    logger.info(f"userspace={userspace}")
+    userspace_dc.create(namespace=name, body=userspace)
+
+
 def create_user_namespace(
     api: client.CoreV1Api,
+    userspace_dc: dynamic.DynamicClient,
     user_name: str,
     user_email: str,
     expected_user_namespaces: Dict[str, str],
     namespaces: List[str],
 ) -> None:
+    env = os.environ.get("ORBIT_ENV", "")
+    if not env:
+        raise ValueError("Orbit Environment ORBIT_ENV is required")
     for team, user_ns in expected_user_namespaces.items():
         try:
             team_namespace = api.read_namespace(name=team).to_dict()
@@ -96,9 +140,13 @@ def create_user_namespace(
             logger.exception("Error retrieving Team Namespace")
             team_uid = None
         if user_ns not in namespaces:
-            logger.info(f"Creating EFS endpoint for {user_ns}...")
-            resp = create_user_efs_endpoint(user=user_name, team_name=team)
-            access_point_id = resp.get("AccessPointId")
+            try:
+                logger.info(f"Creating EFS endpoint for {user_ns}...")
+                efs_ep_resp = create_user_efs_endpoint(user=user_name, team_name=team)
+                access_point_id = efs_ep_resp.get("AccessPointId", "")
+            except Exception as e:
+                logger.warning(f"Error while creating EFS access point for user_name={user_name} and team={team}: {e}")
+                continue
 
             logger.info(f"User namespace {user_ns} doesnt exist. Creating...")
             kwargs = {
@@ -123,10 +171,30 @@ def create_user_namespace(
             body.metadata = client.V1ObjectMeta(**kwargs)
 
             try:
+                # create userspace namespace resource
                 api.create_namespace(body=body)
                 logger.info(f"Created namespace {user_ns}")
-            except ApiException:
-                logger.error(f"Exception when trying to create user namespace {user_ns}")
+            except ApiException as ae:
+                logger.warning(f"Exception when trying to create user namespace {user_ns}")
+                logger.warning(ae.body)
+
+            try:
+                # create userspace custom resource for the given user namespace
+                logger.info(f"Creating userspace custom resource {user_ns}")
+                create_userspace(
+                    userspace_dc=userspace_dc,
+                    name=user_ns,
+                    env=env,
+                    space="user",
+                    team=team,
+                    user=user_name,
+                    user_efsapid=access_point_id,
+                    user_email=user_email,
+                )
+                logger.info(f"Created userspace custom resource {user_ns}")
+            except ApiException as ae:
+                logger.warning(f"Exception when trying to create userspace custom resource {user_ns}")
+                logger.warning(ae.body)
 
 
 def delete_user_efs_endpoint(user_name: str, user_namespace: str, api: client.CoreV1Api) -> None:
@@ -152,32 +220,39 @@ def delete_user_efs_endpoint(user_name: str, user_namespace: str, api: client.Co
 
 
 def delete_user_namespace(
-    api: client.CoreV1Api, user_name: str, expected_user_namespaces: Dict[str, str], namespaces: List[str]
+    api: client.CoreV1Api,
+    userspace_dc: dynamic.DynamicClient,
+    user_name: str,
+    expected_user_namespaces: Dict[str, str],
+    namespaces: List[str],
 ) -> None:
     for user_ns in namespaces:
         if user_ns not in expected_user_namespaces.values():
             delete_user_profile(user_profile=user_ns)
-
             delete_user_efs_endpoint(user_name=user_name, user_namespace=user_ns, api=api)
-
             logger.info(f"User {user_name} is not expected to be part of the {user_ns} namespace. Removing...")
-
+            try:
+                userspace_dc.delete(name=user_ns, namespace=user_ns)
+                logger.info(f"Removed userspace custom resource {user_ns}")
+            except ApiException as ae:
+                logger.warning(f"Exception when trying to remove userspace custom resource {user_ns}")
+                logger.warning(ae.body)
             try:
                 api.delete_collection_namespaced_pod(namespace=user_ns, grace_period_seconds=0)
                 api.delete_namespace(name=user_ns)
                 logger.info(f"Removed namespace {user_ns}")
-            except ApiException:
-                logger.error(f"Exception when trying to remove user namespace {user_ns}")
+            except ApiException as ae:
+                logger.warning(f"Exception when trying to remove user namespace {user_ns}")
+                logger.warning(ae.body)
 
 
 def delete_user_profile(user_profile: str) -> None:
     logger.info(f"Removing profile {user_profile}")
     run_command(f"kubectl delete profile {user_profile} --kubeconfig {KUBECONFIG_PATH}")
-
     time.sleep(5)
 
 
-def manage_user_namespace(event: Dict[str, Any], api: client.CoreV1Api) -> None:
+def manage_user_namespace(event: Dict[str, Any], api: client.CoreV1Api, userspace_dc: dynamic.DynamicClient) -> None:
     user_name = cast(str, event.get("user_name"))
     user_email = cast(str, event.get("user_email"))
     expected_user_namespaces = cast(Dict[str, str], event.get("expected_user_namespaces"))
@@ -191,6 +266,7 @@ def manage_user_namespace(event: Dict[str, Any], api: client.CoreV1Api) -> None:
 
     create_user_namespace(
         api=api,
+        userspace_dc=userspace_dc,
         user_name=user_name,
         user_email=user_email,
         expected_user_namespaces=expected_user_namespaces,
@@ -199,6 +275,7 @@ def manage_user_namespace(event: Dict[str, Any], api: client.CoreV1Api) -> None:
 
     delete_user_namespace(
         api=api,
+        userspace_dc=userspace_dc,
         user_name=user_name,
         expected_user_namespaces=expected_user_namespaces,
         namespaces=all_ns,