+++ CHANGELOG.md
@@ -26,9 +26,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fixing lake-creator pytest shell script and refactoring lake-creator pytest files
 - Removing lake user specific cron job notebooks from regression
 - Removing profiles section from JL UI 
+- Removing lake user specific cron job notebooks from regression 
+- Refactored namespace controller into userspace operator
+- adding userspace cr creation into post auth lambda
 
 ### **Removed**
 
+- removed namespace controller deployment
 
 ## **[1.3.0]**